requirements.txt
anyio==4.8.0
asgiref==3.8.1
astroid==3.0.2
attrs==25.1.0
bcrypt==4.0.1
black==23.12.1
blinker==1.6.2
cattrs==24.1.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
contourpy==1.3.0
cryptography==41.0.1
cycler==0.12.1
dill==0.3.7
distlib==0.3.8
dj-database-url==2.1.0
Django==5.2.1
django-crispy-forms==2.1
djangorestframework==3.16.0
dnspython==2.7.0
espn-api==0.35.0
Faker==12.0.1
filelock==3.13.1
Flask==2.3.2
flatbuffers==24.3.25
fonttools==4.53.1
greenlet==2.0.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.6
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.7
MarkupSafe==2.1.2
matplotlib==3.9.2
mccabe==0.7.0
mixer==7.2.2
mongoengine==0.29.1
mypy-extensions==1.0.0
numpy==2.1.1
openmeteo_requests==1.3.0
openmeteo_sdk==1.18.6
packaging==23.2
pandas==2.2.3
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
pycparser==2.21
pygame==2.4.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pymongo==4.12.1
pyparsing==3.1.4
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.0
pytube==15.0.0
pytz==2023.3.post1
requests==2.31.0
requests-cache==1.2.1
retry-requests==2.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.12.3
typing_extensions==4.6.3
tzdata==2025.2
url-normalize==1.4.3
urllib3==2.1.0
virtualenv==20.25.0
Werkzeug==2.3.4
